+++ muddler_project/src/triggers/Levi_Ataxia/Waking_Up.lua
@@ -1,0 +1,9 @@
+local name = matches[2]
+
+if isTargeted(name) then
+  erAff("sleep")
+  selectString(line, 1)
+  fg("NavajoWhite")
+  resetFormat()
+  targetIshere = true
+end